Core: flattened and split colors to palettes + theme
