+++ packages/core/src/goldTheme.ts
@@ -1,0 +1,27 @@
+// Default theme from GOLD.
+export const goldTheme = {
+	lightForegroundText: '#313131',
+	lightForegroundAction: '#00698f',
+	lightForegroundFocus: '#9263de',
+	lightForegroundMuted: '#626262',
+	lightBackgroundBody: '#FFFFFF',
+	lightBackgroundShade: '#f5f5f5',
+	lightBackgroundBodyAlt: '#ebebeb',
+	lightBackgroundShadeAlt: '#e0e0e0',
+	lightBorder: '#808080',
+	darkForegroundText: '#FFFFFF',
+	darkForegroundAction: '#61daff',
+	darkForegroundFocus: '#c390f9',
+	darkForegroundMuted: '#bdd2d7',
+	darkBackgroundBody: '#135e70',
+	darkBackgroundShade: '#104f5f',
+	darkBackgroundBodyAlt: '#0d414d',
+	darkBackgroundShadeAlt: '#0a323c',
+	darkBorder: '#95b7bf',
+	// system colors
+	accent: undefined, // accent does not exist in GOLD.
+	error: '#d60000',
+	success: '#0b996c',
+	warning: '#f69900',
+	info: '#00bfe9',
+} as const;
