+++ packages/core/src/index.ts
@@ -1,6 +1,8 @@
 export * from './Core';
 export * from './tokens';
 export * from './utils';
-export * from './colors';
+export * from './theme';
+export * from './boxPalette';
+export * from './globalPalette';
 export * from './responsive';
 export * from './context';
